fix conflicts with package renaming
